Feat: 검색 후 입력창 닫히게, 이름 변경
